+++ src/test/java/com/example/teamcity/api/BuildTypeTest.java
@@ -1,7 +1,10 @@
 package com.example.teamcity.api;
 
 import com.example.teamcity.api.enums.RoleTypes;
-import com.example.teamcity.api.models.*;
+import com.example.teamcity.api.models.BuildType;
+import com.example.teamcity.api.models.Project;
+import com.example.teamcity.api.models.Roles;
+import com.example.teamcity.api.models.TestData;
 import com.example.teamcity.api.requests.CheckedRequests;
 import com.example.teamcity.api.requests.unchecked.UncheckedBase;
 import com.example.teamcity.api.spec.Specifications;
@@ -11,7 +14,6 @@ import org.testng.annotations.Test;
 
 import java.util.Arrays;
 
-import static com.example.teamcity.api.enums.BuildStatuses.SUCCESS;
 import static com.example.teamcity.api.enums.Endpoint.*;
 import static com.example.teamcity.api.generators.TestDataGenerator.generate;
 
@@ -78,31 +80,5 @@ public class BuildTypeTest extends BaseApiTest {
                 .create(testData.getBuildType())
                 .then().assertThat().statusCode(HttpStatus.SC_FORBIDDEN)
                 .body(Matchers.containsString("You do not have enough permissions to edit project with id: %s".formatted(testData2.getProject().getId())));
-    }
-
-    @Test(description = "Build runs successfully", groups = {"Positive", "CRUD"})
-    public void buildRunsSuccessfully() throws InterruptedException {
-        // set step with echo "Hello, world!"
-        testData.getBuildType().getSteps().getStep().get(0).setProperties(Properties.builder()
-                .property(Arrays.asList(
-                        Property.builder().name("script.content").value("echo \"Hello, world!\"").build(),
-                        Property.builder().name("use.custom.script").value("true").build()
-                )).build());
-
-        // create project and build type
-        superUserCheckRequests.getRequest(PROJECTS).create(testData.getProject());
-        superUserCheckRequests.getRequest(BUILD_TYPES).create(testData.getBuildType());
-
-        // run build
-        var createdBuild = superUserCheckRequests.<Build>getRequest(BUILD_QUEUE).create(testData.getBuildQueue());
-
-        // wait for build finished
-        for (int i = 0; i < 10; i++) {
-            createdBuild = superUserCheckRequests.<Build>getRequest(BUILDS).read(createdBuild.getId());
-            if (createdBuild.getStatus() != null && createdBuild.getStatus().equals(SUCCESS))
-                break;
-            Thread.sleep(1000);
-        }
-        softy.assertEquals(createdBuild.getStatus(), SUCCESS, "Incorrect build status %s".formatted(createdBuild.getStatus()));
     }
 }
